Discord event syncing (#45)
* Sync website events with discord

Making a event in the webui also makes one in the discord. Deleting also
deletes the discord event. Todo: updating the event on the website also
updates the discord event

* Updating event in the webUI updates discord event

maybe todo: when updating an event and removing the share on discord
option this should delete the discord event, but i though that might be
a bit confusing, and maybe not that usefull

* accidentally removed joi import last commit
* forgot to run prettier
* rename function name to english, and remove bind:value={*} on checkboxes as it is redundant
* move discord event logic from utils to discord
* remove unused imports